Enhance Unit Tests for HTTP Metrics and Device Service
- Updated `test_http_metrics.c` to introduce new constants for response sizes, latencies, and iteration counts, improving test clarity and maintainability.
- Refactored test cases to utilize defined constants, ensuring consistency and reducing magic numbers in assertions.
- Enhanced `test_device_service.c` by adding new constants for operation name lengths and service counts, improving readability and maintainability of test data structures.
- Adjusted mock response handling to use defined constants for SOAP envelope lengths, ensuring compliance with project standards.

These changes strengthen the unit testing framework for HTTP metrics and device service, aligning with project goals for improved test reliability and maintainability.

+++ cross-compile/onvif/tests/src/unit/utils/test_memory_utils.c
@@ -18,7 +18,7 @@
 #define TEST_ALLOC_SIZE           1024
 #define TEST_ALLOC_LAST_INDEX     (TEST_ALLOC_SIZE - 1)
 #define TEST_SMALL_ALLOC_SIZE     512
-#define TEST_LARGE_ALLOC_SIZE     (1024 * 1024)
+#define TEST_LARGE_ALLOC_SIZE     ((size_t)1024 * 1024)
 #define TEST_MEDIUM_ALLOC_SIZE_1  100
 #define TEST_MEDIUM_ALLOC_SIZE_2  200
 #define TEST_ITERATION_ALLOC_SIZE 64
@@ -231,3 +231,4 @@ void test_unit_dynamic_buffer(void** state) {
 }
 
 // Test functions are called from test_runner.c
+// Test comment
